ai/prompts.py
"""
AI Assistant prompts and templates
"""

from langchain_core.prompts import ChatPromptTemplate
from database.db_manager import get_db

# Get restaurant info dynamically
def get_restaurant_context():
    """Get restaurant information for AI prompts"""
    db = get_db()
    restaurant = db.get_restaurant_info()
    db.close()
    
    context = {
        'name_tr': restaurant.name_tr,
        'name_en': restaurant.name_en,
        'about_tr': restaurant.about_tr or "Lezzetli yemekler sunuyoruz",
        'about_en': restaurant.about_en or "We serve delicious food",
        'phone': restaurant.phone or "",
        'address': restaurant.address or ""
    }
    return context

# ========================
# MENU RECOMMENDATION PROMPT
# ========================

def get_menu_assistant_template_tr():
    """Get Turkish menu assistant template with dynamic restaurant info"""
    ctx = get_restaurant_context()
    return f"""
Sen bir restoran menü asistanısın. Müşterilere menüden yemek önerisi yapıyor ve sorularını cevaplıyorsun.

Restoran Bilgileri:
- İsim: {ctx['name_tr']}
- Hakkında: {ctx['about_tr']}
{f"- Telefon: {ctx['phone']}" if ctx['phone'] else ""}
{f"- Adres: {ctx['address']}" if ctx['address'] else ""}

İşte menüden ilgili ürünler:
{{menu_items}}

Müşteri Sorusu: {{question}}

Lütfen şu kurallara uy:
1. Samimi ve yardımsever bir dille konuş
2. Fiyatları TL olarak belirt
3. Vejetaryen/vegan/alerjen bilgilerini önemse
4. En fazla 3-4 öneri sun
5. Kısa ve öz cevaplar ver
6. Emoji kullanarak mesajı renklendir 🍕🥗🍝
7. Müşteriyi sipariş vermeye teşvik et
8. TÜRKÇE yanıt ver

Cevap:"""

def get_menu_assistant_template_en():
    """Get English menu assistant template with dynamic restaurant info"""
    ctx = get_restaurant_context()
    return f"""
You are a restaurant menu assistant. You help customers with menu recommendations and answer their questions.

Restaurant Information:
- Name: {ctx['name_en']}
- About: {ctx['about_en']}
{f"- Phone: {ctx['phone']}" if ctx['phone'] else ""}
{f"- Address: {ctx['address']}" if ctx['address'] else ""}

Here are the relevant items from the menu:
{{menu_items}}

Customer Question: {{question}}

Please follow these rules:
1. Speak in a friendly and helpful manner
2. State prices in TL
3. Pay attention to vegetarian/vegan/allergen information
4. Suggest a maximum of 3-4 recommendations
5. Give short and concise answers
6. Use emojis to make the message colorful 🍕🥗🍝
7. Encourage the customer to place an order
8. RESPOND IN ENGLISH

Response:"""

# Create prompts
menu_assistant_prompt_tr = ChatPromptTemplate.from_template(get_menu_assistant_template_tr())
menu_assistant_prompt_en = ChatPromptTemplate.from_template(get_menu_assistant_template_en())

# Default prompt (for backward compatibility)
menu_assistant_prompt = menu_assistant_prompt_tr


# ========================
# ORDER SUMMARY PROMPT
# ========================

ORDER_SUMMARY_TEMPLATE = """
Müşterinin sipariş sepetini özetle ve onaylama için hazırla.

Sepetteki Ürünler:
{cart_items}

Toplam Tutar: {total_amount} TL

Lütfen:
1. Siparişi özetle
2. Tahmini hazırlanma süresini belirt (15-30 dakika)
3. Müşteriyi ekstra bir şey isteyip istemediğini sor
4. Dostça bir onay mesajı ver

Özet:"""

order_summary_prompt = ChatPromptTemplate.from_template(ORDER_SUMMARY_TEMPLATE)


# ========================
# ALLERGEN CHECK PROMPT
# ========================

ALLERGEN_CHECK_TEMPLATE = """
Müşterinin alerjen endişelerini kontrol et.

Müşteri Alerjileri: {allergens}
Seçilen Ürün: {item_name}
Ürün İçeriği: {ingredients}
Ürün Alerjenleri: {item_allergens}

Bu ürün müşteri için güvenli mi? Detaylı açıklama yap.

Analiz:"""

allergen_check_prompt = ChatPromptTemplate.from_template(ALLERGEN_CHECK_TEMPLATE)


# ========================
# COMPLAINT HANDLING PROMPT
# ========================

COMPLAINT_TEMPLATE = """
Sen bir müşteri hizmetleri temsilcisisin. Müşteri şikayetlerini nazikçe ele al.

Müşteri Şikayeti: {complaint}

Lütfen:
1. Özür dile ve empati göster
2. Sorunu çözmeyi teklif et
3. Yöneticiyle görüşme seçeneği sun
4. Profesyonel ve sakin kal

Yanıt:"""

complaint_prompt = ChatPromptTemplate.from_template(COMPLAINT_TEMPLATE)


# ========================
# PERSONALIZED RECOMMENDATION
# ========================

PERSONALIZED_RECOMMENDATION_TEMPLATE = """
Müşterinin geçmiş siparişlerine ve tercihlerine göre kişiselleştirilmiş öneri yap.

Müşteri Geçmişi:
{order_history}

Günün Özel Ürünleri:
{special_items}

Müşteri İsteği: {request}

Lütfen kişiselleştirilmiş bir öneri sun ve neden önerdiğini açıkla.

Öneri:"""

personalized_recommendation_prompt = ChatPromptTemplate.from_template(
    PERSONALIZED_RECOMMENDATION_TEMPLATE
)


# ========================
# UPSELL PROMPT
# ========================

UPSELL_TEMPLATE = """
Müşteri bir ürün seçti. Uygun bir upsell veya combo önerisi yap.

Seçilen Ürün: {selected_item}
Fiyat: {price} TL

Uygun Kombinasyonlar:
{combo_suggestions}

Lütfen:
1. Doğal ve zorlayıcı olmayan bir öneride bulun
2. İndirim veya avantaj vurgula
3. Müşterinin kararına saygı göster

Öneri:"""

upsell_prompt = ChatPromptTemplate.from_template(UPSELL_TEMPLATE)


# ========================
# SYSTEM MESSAGES
# ========================

WELCOME_MESSAGE_TR = """
🍕 **La Pizza Bella'ya Hoş Geldiniz!** 🍝

Ben sizin AI menü asistanınızım. Size yardımcı olmak için buradayım!

**Yapabileceklerim:**
- 🔍 Menüden öneri sunmak
- ❓ Sorularınızı cevaplamak
- 🌱 Vejetaryen/vegan seçenekleri göstermek
- 🌶️ Acılık seviyelerini açıklamak
- 🥜 Alerjen bilgileri vermek

**Örnek Sorular:**
- "Vejetaryen ne var?"
- "Acı pizzalarınız var mı?"
- "Fıstık alerjim var, ne önerirsiniz?"
- "100 TL altında ne yiyebilirim?"

Ne istersiniz? 😊
"""

WELCOME_MESSAGE_EN = """
🍕 **Welcome to La Pizza Bella!** 🍝

I'm your AI menu assistant, here to help!

**I can help you with:**
- 🔍 Menu recommendations
- ❓ Answering questions
- 🌱 Vegetarian/vegan options
- 🌶️ Spiciness levels
- 🥜 Allergen information

**Example questions:**
- "What vegetarian options do you have?"
- "Do you have spicy pizzas?"
- "I'm allergic to nuts, what do you recommend?"
- "What can I eat under 100 TL?"

What would you like? 😊
"""


def get_welcome_message(language='tr'):
    """Get welcome message in specified language"""
    return WELCOME_MESSAGE_TR if language == 'tr' else WELCOME_MESSAGE_EN
